Dropping support for external links
After introducing them, I noticed that they're not really useful. Of
course, they allow me to provide my readers with links to articles or
other things, but in the end, it just encourages me to post too much of
those kind of links.

This is where Twitter comes in. I usually post most of the linked stuff
there, so this will be the right place to continue it.