Better footnotes
If the client has JavaScript enabled, they're now shown on the right
side of the screen as soon as the user clicks on a footnote link.

That means it's way faster to get to the content now. The user also
doesn't need to scroll around to view the original note.